requirements.txt
picamera2>=0.3.12
opencv-python>=4.8.0
numpy>=1.24.0
mediapipe>=0.10.8
dataclasses>=0.6
RPi.GPIO>=0.7.0
vosk>=0.3.45
pyaudio>=0.2.13 